+++ packages/lib/src/types/api/models/balanceAccounts.ts
@@ -1,5 +1,5 @@
-import { Schema } from './api/utils';
-import { components } from './openapi/BalanceAccountsResource';
+import { Schema } from '../utils';
+import { components } from '../resources/BalanceAccountsResource';
 
 export type IBalanceAccountBase = Schema<components, 'BalanceAccountBase'>;
 export type IBalance = Schema<components, 'Balance'>;
