+++ packages/lib/src/types/api/resources/PingResource.d.ts
