+++ packages/lib/src/types/api/resources/SetupResource.d.ts
