+++ packages/lib/src/types/api/resources/TransactionsResource.d.ts
