--- packages/lib/src/types/models/api/index.ts
@@ -1,2 +1,0 @@
-export * from '../transactions';
-export * from '../balanceAccount';
